fix: id in domain changed to int64, matching monitor

+++ services/erp-service/internal/domain/order_report.go
@@ -24,7 +24,7 @@ const (
 type OrderReport struct {
 	ID                   int64
 	CompanyID            string
-	OrderOperationID     int // referenced by OrderOperation id, (Not order)
+	OrderOperationID     int64 // referenced by OrderOperation id, (Not order)
 	ProductionResourceID string
 	Quantity             float64
 	RestQuantity         float64
